Moved getting current project and getting quartus bin path to path utils
